Update patch set 10
Patch Set 10:

This is an abuse of ACLs.
Initial package reviews require that you send a formal report on bugzilla to comply with legal requirements which should be signed off by a core packager. Besides this is made to ensure proper quality of the packaging. There are things that CI do not test like is the package complying with all guidelines? is it properly running? etc.

If you have any questions about the reviewing process, please come and ask us.

Patch-set: 10